Remove test profile 'giflw' (duplicate of 'bvfalcon')
This profile was an exact copy before of used plugin repository, group
ID and version, when it was still referencing BinTray. After I upgraded
it, it is still a duplicate, I just noticed. So instead of running the
exact same test twice, always passing or failing with the same results,
remove the duplicate.